+++ src/root.zig
@@ -35,6 +35,7 @@ pub const analysis = @import("analysis.zig");
 pub const stream_writer = @import("stream_writer.zig");
 pub const toml_parser = @import("toml_parser.zig");
 pub const toml_classifier_parser = @import("toml_classifier_parser.zig");
+pub const dcd = @import("dcd.zig");
 
 test {
     _ = shrake_rupley;
@@ -48,4 +49,5 @@ test {
     _ = stream_writer;
     _ = toml_parser;
     _ = toml_classifier_parser;
+    _ = dcd;
 }